requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
python-dotenv==1.0.1
openai==1.59.7
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
Jinja2==3.1.4
